+++ CHANGELOG.md
@@ -8,6 +8,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- `telemetry.distro.name` resource attribute with the value `splunk-otel-go`.
+- `telemetry.distro.version` resource attribute with the value of the current
+  release version.
+
+### Deprecated
+
+- `splunk.distro.version` resource attribute
+
 ## [1.25.0] - 2025-03-20
 
 This release adds experimental logs support. Set `OTEL_LOGS_EXPORTER=otlp` to